implememnted the part for 30 credits, I don't knwo if it works but at least it's not broken, nodes for some reason don' t correctly accumulate their balance

+++ src/include/common.h
@@ -100,7 +100,7 @@ struct parameters
 typedef struct user_t
 {
     pid_t pid;
-    int balance;
+    unsigned int balance;
     enum
     {
         alive,
@@ -112,7 +112,7 @@ typedef struct user_t
 typedef struct node_t
 {
     pid_t pid;
-    int balance;
+    unsigned long balance;
     enum
     {
         available,
